java-based configuration: @Import, @ImportResource
Use @Import for importing the configuration from another module.
