+++ src/main/java/org/stanislav/spring/bfpp/LogBeanFactoryPostProcessor.java
@@ -1,16 +1,17 @@
-package org.stanislav.bfpp;
+package org.stanislav.spring.bfpp;
 
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
 import org.springframework.core.Ordered;
+import org.springframework.core.PriorityOrdered;
 import org.springframework.stereotype.Component;
 
 /**
  * @author Stanislav Hlova
  */
 @Component
-public class VerifyBreakpointBeanFactoryPostProcessor implements BeanFactoryPostProcessor, Ordered {
+public class LogBeanFactoryPostProcessor implements BeanFactoryPostProcessor, PriorityOrdered {
     @Override
     public void postProcessBeanFactory(ConfigurableListableBeanFactory beanFactory) throws BeansException {
         System.out.println();
@@ -18,6 +19,6 @@ public class VerifyBreakpointBeanFactoryPostProcessor implements BeanFactoryPost
 
     @Override
     public int getOrder() {
-        return 0;
+        return Ordered.HIGHEST_PRECEDENCE;
     }
 }
