+++ src/main/java/org/stanislav/spring/service/CompanyService.java
@@ -1,4 +1,4 @@
-package org.stanislav.service;
+package org.stanislav.spring.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
