+++ src/main/java/org/stanislav/web/config/WebConfiguration.java
@@ -1,0 +1,10 @@
+package org.stanislav.web.config;
+
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * @author Stanislav Hlova
+ */
+@Configuration
+public class WebConfiguration {
+}
